requirements.txt
pandas~=1.5.3
numpy~=1.24.3
pytz~=2022.7
apache-airflow[postgres]
apache-airflow-providers-google
